some changes were made
