+++ src/Components/login/materialStyles.js
@@ -1,26 +1,23 @@
-const textFieldBorderStyle = {
-  "& .MuiInputLabel-root": {},
-  borderBottom: "1px solid rgb(179, 175, 177)",
-};
-
-const userExistsMessageStyle = {
-  text: "The user already exists. Log In.",
-  messageColor: { color: "rgb(237, 85, 151)" },
-};
-
-const messagePassWordNotMatchStyles = {
-  text: "Passwords do not match.",
-  messageColor: { color: "rgb(237, 85, 151)" },
-};
-
-const messageSignupStyles = {
+const messageSignInStyle = {
   text: "Enter your detalls below.",
   messageColor: { color: "rgb(150, 157, 166)" },
+  borderBottonStyle: { borderBottom: "1px solid rgb(179, 175, 177)" },
+};
+
+const failedLoginMessageStyle = {
+  text: "Oops! That email or password combination is not valid.",
+  messageColor: { color: "rgb(237, 85, 151)" },
+  borderBottonStyle: { borderBottom: "1px solid rgb(237, 85, 151)" },
+};
+
+const UserDoesNotExistsMessageStyle = {
+  text: "Oops! Username does not exist",
+  messageColor: { color: "rgb(237, 85, 151)" },
+  borderBottonStyle: { borderBottom: "1px solid rgb(237, 85, 151)" },
 };
 
 export {
-  textFieldBorderStyle,
-  userExistsMessageStyle,
-  messagePassWordNotMatchStyles,
-  messageSignupStyles,
+  messageSignInStyle,
+  failedLoginMessageStyle,
+  UserDoesNotExistsMessageStyle,
 };
