--- src/Components/signUp/materialStylesLogin.js
@@ -1,23 +1,0 @@
-const messageSignInStyle = {
-  text: "Enter your detalls below.",
-  messageColor: { color: "rgb(150, 157, 166)" },
-  borderBottonStyle: { borderBottom: "1px solid rgb(179, 175, 177)" },
-};
-
-const failedLoginMessageStyle = {
-  text: "Oops! That email or password combination is not valid.",
-  messageColor: { color: "rgb(237, 85, 151)" },
-  borderBottonStyle: { borderBottom: "1px solid rgb(237, 85, 151)" },
-};
-
-const UserDoesNotExistsMessageStyle = {
-  text: "Oops! Username does not exist",
-  messageColor: { color: "rgb(237, 85, 151)" },
-  borderBottonStyle: { borderBottom: "1px solid rgb(237, 85, 151)" },
-};
-
-export {
-  messageSignInStyle,
-  failedLoginMessageStyle,
-  UserDoesNotExistsMessageStyle,
-};
