+++ src/Components/signUp/materialStylesSignUp.js
@@ -1,0 +1,26 @@
+const textFieldBorderStyle = {
+  "& .MuiInputLabel-root": {},
+  borderBottom: "1px solid rgb(179, 175, 177)",
+};
+
+const userExistsMessageStyle = {
+  text: "The user already exists. Log In.",
+  messageColor: { color: "rgb(237, 85, 151)" },
+};
+
+const messagePassWordNotMatchStyles = {
+  text: "Passwords do not match.",
+  messageColor: { color: "rgb(237, 85, 151)" },
+};
+
+const messageSignupStyles = {
+  text: "Enter your detalls below.",
+  messageColor: { color: "rgb(150, 157, 166)" },
+};
+
+export {
+  textFieldBorderStyle,
+  userExistsMessageStyle,
+  messagePassWordNotMatchStyles,
+  messageSignupStyles,
+};
